% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/peruseR.R
\name{peruseR}
\alias{peruseR}
\title{Resource for proposing and planning useR! conferences}
\usage{
peruseR()
}
\description{
\code{peruseR} is a collection of document templates and data that we hope
will be useful for future planners of useR! conferences. It includes LaTeX
document templates as well as examples of previous year proposals that may
be used to generate a new proposal.

At the moment, not too much.
}

